+++ Tests/AppLogOutputChecker.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 98af7d4f91f854149b88a674aff09b12
+guid: e0cb6bc08dd98eb48998d487a0c6a0b1
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
