+++ Tests/AppLogWriter.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: cd96cfa7fe278464fb40c84f5d78925c
+guid: a421165f14c64e545b69a5d6ecce5e26
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
