--- Tests/DefaultLogTest.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8078e98a781d07343a7c53b408a96358
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
