+++ Tests/LogOutputChekerTest.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 6a2bd9f78e05e6945a3fd1c15b9ae986
+guid: 47806aa6ecbdd4649bf26174f9ac02d9
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
