+++ Tests/LoggingManagerTest.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 03357914ef5e15f4692b5e4f4be6485b
+guid: 553141693b9b55944808a4b1eee3ddd4
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
